requirements.txt
django==3.2.6
pymysql==1.0.2
django-environ==0.4.5
gunicorn==20.1.0
requests==2.26.0
whitenoise==5.3.0
django-cors-headers==3.10.1
django-heroku==0.3.1
linkedin_api~=2.0.0a1
waitress==2.0.0
bs4==0.0.1
django-storages==1.12.3
boto3==1.20.26